LPS-66937 let's keep these nicely separated and move them to a different package
