+++ modules/apps/web-experience/journal/journal-service/src/main/resources/portlet.properties
@@ -119,7 +119,7 @@ journal.folder.icon.check.count=true
 # Specify the strategy used when Journal content is imported using the LAR
 # system.
 #
-lar.creation.strategy=com.liferay.journal.exportimport.data.handler.JournalCreationStrategyImpl
+lar.creation.strategy=com.liferay.journal.internal.exportimport.creation.strategy.JournalCreationStrategyImpl
 
 #
 # Set this to true if journal articles should be published to live by default.
